+++ src/components/SearchFilters/SearchFilters.js
@@ -1,10 +1,13 @@
 import { useCallback } from 'react';
 import PropTypes from 'prop-types';
-import {
-  FormattedMessage,
-} from 'react-intl';
+import { useIntl } from 'react-intl';
 import omit from 'lodash/omit';
 
+import {
+  Accordion,
+  Checkbox,
+  FilterAccordionHeader,
+} from '@folio/stripes/components';
 import {
   AcqDateRangeFilter,
 } from '@folio/stripes-acq-components';
@@ -21,9 +24,12 @@ const DATE_FORMAT = 'YYYY-MM-DD';
 
 const propTypes = {
   activeFilters: PropTypes.object.isRequired,
+  applyExcludeSeeFromLimiter: PropTypes.func.isRequired,
+  isExcludedSeeFromLimiter: PropTypes.bool.isRequired,
   isSearching: PropTypes.bool.isRequired,
   query: PropTypes.string.isRequired,
   setFilters: PropTypes.func.isRequired,
+  setIsExcludedSeeFromLimiter: PropTypes.func.isRequired,
 };
 
 const SearchFilters = ({
@@ -31,7 +37,12 @@ const SearchFilters = ({
   isSearching,
   setFilters,
   query,
+  isExcludedSeeFromLimiter,
+  setIsExcludedSeeFromLimiter,
+  applyExcludeSeeFromLimiter,
 }) => {
+  const intl = useIntl();
+
   const [filterAccordions, { handleSectionToggle }] = useSectionToggle({
     [FACETS.HEADING_TYPE]: false,
   });
@@ -50,6 +61,7 @@ const SearchFilters = ({
         [name]: values,
       };
     });
+  // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   const onClearFilter = (filter) => {
@@ -58,9 +70,23 @@ const SearchFilters = ({
 
   return (
     <>
+      <Accordion
+        closedByDefault
+        displayClearButton={isExcludedSeeFromLimiter}
+        header={FilterAccordionHeader}
+        label={intl.formatMessage({ id: 'ui-marc-authorities.search.references' })}
+        onClearFilter={() => setIsExcludedSeeFromLimiter(false)}
+      >
+        <Checkbox
+          label={intl.formatMessage({ id: 'ui-marc-authorities.search.excludeSeeFrom' })}
+          onChange={applyExcludeSeeFromLimiter}
+          checked={isExcludedSeeFromLimiter}
+        />
+      </Accordion>
+
       <MultiSelectionFacet
         id={FACETS.HEADING_TYPE}
-        label={<FormattedMessage id={`ui-marc-authorities.filters.${FACETS.HEADING_TYPE}`} />}
+        label={intl.formatMessage({ id: `ui-marc-authorities.search.${FACETS.HEADING_TYPE}` })}
         name={FACETS.HEADING_TYPE}
         open={filterAccordions[FACETS.HEADING_TYPE]}
         options={facets[FACETS.HEADING_TYPE]?.values || []}