+++ src/queries/utils/buildQuery.js
@@ -1,9 +1,14 @@
+import { upperFirst } from 'lodash';
+
 import {
   searchableIndexesValues,
   searchableIndexesMap,
 } from '../../constants';
 
-const buildQuery = (searchIndex) => {
+const buildQuery = ({
+  searchIndex,
+  isExcludedSeeFromLimiter = false,
+}) => {
   const indexData = searchableIndexesMap[searchIndex || searchableIndexesValues.KEYWORD];
 
   const queryStrings = indexData.map(data => {
@@ -11,40 +16,40 @@ const buildQuery = (searchIndex) => {
 
     const queryTemplate = name => `${name}=="%{query}"`;
 
-    const capitalizeFirstLetter = ([first, ...rest]) => first.toUpperCase() + rest.join('');
-
     if (data.plain) {
       const query = queryTemplate(data.name);
 
       queryParts.push(query);
     }
 
-    if ((data.sft || data.saft) && data.plain) {
-      const name = capitalizeFirstLetter(data.name);
+    if (!isExcludedSeeFromLimiter) {
+      if ((data.sft || data.saft) && data.plain) {
+        const name = upperFirst(data.name);
 
-      if (data.sft) {
-        const query = queryTemplate(`sft${name}`);
+        if (data.sft) {
+          const query = queryTemplate(`sft${name}`);
+
+          queryParts.push(query);
+        }
+
+        if (data.saft) {
+          const query = queryTemplate(`saft${name}`);
+
+          queryParts.push(query);
+        }
+      }
+
+      if (data.sft && !data.plain) {
+        const query = queryTemplate(data.name);
 
         queryParts.push(query);
       }
 
-      if (data.saft) {
-        const query = queryTemplate(`saft${name}`);
+      if (data.saft && !data.plain) {
+        const query = queryTemplate(data.name);
 
         queryParts.push(query);
       }
-    }
-
-    if (data.sft && !data.plain) {
-      const query = queryTemplate(data.name);
-
-      queryParts.push(query);
-    }
-
-    if (data.saft && !data.plain) {
-      const query = queryTemplate(data.name);
-
-      queryParts.push(query);
     }
 
     return queryParts;