+++ src/queries/utils/buildQuery.test.js
@@ -4,7 +4,10 @@ import { searchableIndexesValues } from '../../constants';
 describe('Given buildQuery', () => {
   describe('when index without any prefix provided', () => {
     it('should return correct query', () => {
-      const query = buildQuery(searchableIndexesValues.IDENTIFIER);
+      const query = buildQuery({
+        searchIndex: searchableIndexesValues.IDENTIFIER,
+        isExcludedSeeFromLimiter: true,
+      });
 
       expect(query).toBe('(identifier=="%{query}")');
     });
@@ -12,7 +15,9 @@ describe('Given buildQuery', () => {
 
   describe('when index with plain, sft, and saft prefixes provided', () => {
     it('should return correct query', () => {
-      const query = buildQuery(searchableIndexesValues.PERSONAL_NAME);
+      const query = buildQuery({
+        searchIndex: searchableIndexesValues.PERSONAL_NAME,
+      });
 
       expect(query).toBe('(personalName=="%{query}" or sftPersonalName=="%{query}" or saftPersonalName=="%{query}")');
     });
@@ -20,7 +25,9 @@ describe('Given buildQuery', () => {
 
   describe('when index with different names for plain, sft, and saft prefixes provided', () => {
     it('should return correct query', () => {
-      const query = buildQuery(searchableIndexesValues.GEOGRAPHIC_NAME);
+      const query = buildQuery({
+        searchIndex: searchableIndexesValues.GEOGRAPHIC_NAME,
+      });
 
       expect(query).toBe('(geographicName=="%{query}" or sftGeographicTerm=="%{query}" or saftGeographicTerm=="%{query}")');
     });
@@ -28,7 +35,9 @@ describe('Given buildQuery', () => {
 
   describe('when keyword index provided', () => {
     it('should return correct query for keyword index', () => {
-      const query = buildQuery('');
+      const query = buildQuery({
+        searchIndex: '',
+      });
 
       expect(query).toBe('(keyword=="%{query}")');
     });